review feedback: move functionality into state module
